backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-admin
flask-security
Flask-Login
Flask-RESTful
Flask-HTTPAuth
flask-marshmallow
marshmallow-sqlalchemy
email-validator
dynaconf
psycopg2-binary
python-dotenv